requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.0
fastcrc==0.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==6.0.1
openai==1.102.0
pydantic==2.11.7
pydantic-core==2.33.2
pymavlink==2.4.49
pyqt6==6.9.1
pyqt6-qt6==6.9.1
pyqt6-sip==13.10.2
pyserial==3.5
pyyaml==6.0.2
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.15.0
typing-inspection==0.4.1
